+++ sdk/client/src/listeners/first-load-listeners.tsx
@@ -17,6 +17,7 @@ import {
 	matchPerformanceTimingsWithRequestResponsePair,
 	shouldNetworkRequestBeRecorded,
 } from './network-listener/utils/utils'
+import { shutdown } from '../otel'
 
 // Note: This class is used by both firstload and client. When constructed in client, it will match the current
 // codebase. When constructed in firstload, it will match the codebase at the time the npm package was published.
@@ -138,9 +139,7 @@ export class FirstLoadListeners {
 			),
 		)
 		if (this.options.enableOtelTracing) {
-			import('@highlight-run/client/src/otel').then(({ shutdown }) => {
-				this.listeners.push(shutdown)
-			})
+			this.listeners.push(shutdown)
 		}
 		FirstLoadListeners.setupNetworkListener(this, this.options)
 	}
@@ -177,7 +176,6 @@ export class FirstLoadListeners {
 			sThis.networkHeadersToRedact = []
 			sThis.networkBodyKeysToRedact = []
 			sThis.urlBlocklist = []
-			sThis.networkBodyKeysToRecord = []
 			sThis.networkBodyKeysToRecord = []
 		} else if (typeof options?.networkRecording === 'boolean') {
 			sThis.disableNetworkRecording = !options.networkRecording
@@ -267,8 +265,8 @@ export class FirstLoadListeners {
 					backendUrl: sThis._backendUrl,
 					tracingOrigins: sThis.tracingOrigins,
 					urlBlocklist: sThis.urlBlocklist,
-					sessionSecureID: options.sessionSecureID,
 					bodyKeysToRecord: sThis.networkBodyKeysToRecord,
+					otelEnabled: !!options.enableOtelTracing,
 				}),
 			)
 		}