+++ sdk/client/src/listeners/network-listener/utils/fetch-listener.ts
@@ -13,6 +13,7 @@ import {
 
 import { NetworkListenerCallback } from '../network-listener'
 import { getBodyThatShouldBeRecorded } from './xhr-listener'
+import { getActiveSpan } from '../../../otel'
 
 export interface HighlightFetchWindow extends WindowOrWorkerGlobalScope {
 	_originalFetch: WindowOrWorkerGlobalScope['fetch']
@@ -27,8 +28,9 @@ export const FetchListener = (
 	backendUrl: string,
 	tracingOrigins: boolean | (string | RegExp)[],
 	urlBlocklist: string[],
-	bodyKeysToRedact?: string[],
-	bodyKeysToRecord?: string[],
+	bodyKeysToRedact: string[],
+	bodyKeysToRecord: string[] | undefined,
+	otelEnabled: boolean,
 ) => {
 	const originalFetch = window._fetchProxy
 
@@ -38,7 +40,7 @@ export const FetchListener = (
 			return originalFetch.call(this, input, init)
 		}
 
-		const [sessionSecureID, requestId] = createNetworkRequestId()
+		const [sessionSecureID, requestId] = createNetworkRequestId(otelEnabled)
 		if (shouldNetworkRequestBeTraced(url, tracingOrigins)) {
 			init = init || {}
 			// Pre-existing headers could be one of three different formats; this reads all of them.