+++ sdk/highlight-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @highlight-run/react
 
+## 4.0.1
+
+### Patch Changes
+
+-   Updated dependencies [0067ea6b5]
+    -   highlight.run@9.2.1
+
 ## 4.0.0
 
 ### Patch Changes